% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_database.R
\name{example_database}
\alias{example_database}
\title{Function that generates the test database}
\usage{
example_database(path_date, separator)
}
\arguments{
\item{path_date}{is a directory path containing the csv file}

\item{separator}{is the separator for files in csv format}
}
\value{
This function return is a dataframe with database
}
\description{
The example_database function reads a csv file available on the computer system and returns a dataframe as an example of a database to be used by the functions of the bibliorefer package
}
\examples{

#Call the example_database function

file_db <- system.file("extdata","example_database.csv", package = "gerefer")
separator <- ","
date_sreference <- example_database(file_db, separator)
date_sreference

}
\references{
Aria, M. & Cuccurullo, C. (2017) bibliometrix: An R-tool for comprehensive science mapping analysis, Journal of Informetrics, 11(4), pp 959-975, Elsevier
}
